==> notebooks/01_AnaliseCampos.ipynb <==
import pandas as pd

### Extract Columns List

enem22 = pd.read_csv("../../dados_enem/microdados_enem_2022/DADOS/MICRODADOS_ENEM_2022.csv", sep=';', encoding = "ISO-8859-1")
enem22_clm = enem22.columns.tolist()

print('Shape:\t',enem22.shape,'\nColumns:',enem22.columns)
del enem22

enem21 = pd.read_csv("../../dados_enem/microdados_enem_2021/DADOS/MICRODADOS_ENEM_2021.csv", sep=';', encoding = "ISO-8859-1")
enem21_clm = enem21.columns.tolist()

print('Shape:\t',enem21.shape,'\nColumns:',enem21.columns)
del enem21

enem20 = pd.read_csv("../../dados_enem/microdados_enem_2020/DADOS/MICRODADOS_ENEM_2020.csv", sep=';', encoding = "ISO-8859-1")
enem20_clm = enem20.columns.tolist()

print('Shape:\t',enem20.shape,'\nColumns:',enem20.columns)
del enem20

enem19 = pd.read_csv("../../dados_enem/microdados_enem_2019/DADOS/MICRODADOS_ENEM_2019.csv", sep=';', encoding = "ISO-8859-1")
enem19_clm = enem19.columns.tolist()

print('Shape:\t',enem19.shape,'\nColumns:',enem19.columns)
del enem19

enem18 = pd.read_csv("../../dados_enem/microdados_enem_2018/DADOS/MICRODADOS_ENEM_2018.csv", sep=';', encoding = "ISO-8859-1")
enem18_clm = enem18.columns.tolist()

print('Shape:\t',enem18.shape,'\nColumns:',enem18.columns)
del enem18

enem17 = pd.read_csv("../../dados_enem/microdados_enem_2017/DADOS/MICRODADOS_ENEM_2017.csv", sep=';', encoding = "ISO-8859-1")
enem17_clm = enem17.columns.tolist()

print('Shape:\t',enem17.shape,'\nColumns:',enem17.columns)
del enem17

enem16 = pd.read_csv("../../dados_enem/microdados_enem_2016/DADOS/MICRODADOS_ENEM_2016.csv", sep=';', encoding = "ISO-8859-1")
enem16_clm = enem16.columns.tolist()

print('Shape:\t',enem16.shape,'\nColumns:',enem16.columns)
del enem16

enem15 = pd.read_csv("../../dados_enem/microdados_enem_2015/DADOS/MICRODADOS_ENEM_2015.csv", sep=';', encoding = "ISO-8859-1")
enem15_clm = enem15.columns.tolist()

print('Shape:\t',enem15.shape,'\nColumns:',enem15.columns)
del enem15

enem14 = pd.read_csv("../../dados_enem/microdados_enem_2014/DADOS/MICRODADOS_ENEM_2014.csv", sep=';', encoding = "ISO-8859-1")
enem14_clm = enem14.columns.tolist()

print('Shape:\t',enem14.shape,'\nColumns:',enem14.columns)
del enem14

enem13 = pd.read_csv("../../dados_enem/microdados_enem_2013/DADOS/MICRODADOS_ENEM_2013.csv", sep=';', encoding = "ISO-8859-1")
enem13_clm = enem13.columns.tolist()

print('Shape:\t',enem13.shape,'\nColumns:',enem13.columns)
del enem13

### Analyze Common Columns

columns = [enem13_clm,enem14_clm,enem15_clm,enem16_clm,enem17_clm,enem18_clm,enem19_clm,enem20_clm,enem21_clm,enem22_clm]

intercept = set(columns[0])
for s in columns[1:]:
    intercept.intersection_update(s)

for col in intercept:
    print(col)
