==> notebooks/02_TratamentoDados.ipynb <==
import time
start = time.time()

import os
import re
import pickle
import pandas as pd
from datetime import timedelta
from unicodedata import normalize, combining

from geopy.geocoders import Nominatim
geolocator = Nominatim(user_agent='luizhemelo', timeout=10)

from warnings import simplefilter
simplefilter(action="ignore", category=pd.errors.PerformanceWarning)

PATH = "../../dados_enem"  # path to datasets
cols = ['NO_MUNICIPIO_PROVA', 'CO_MUNICIPIO_PROVA', 'TP_COR_RACA']  # cities col reference

## Funções auxiliares

# city location cache
path = "../data/cities.pickle"
if os.path.exists(path):
    file = open(path, 'rb')
    cities = pickle.load(file)
    file.close()
else:
    cities = {
        'poxereu':[-15.8290891, -54.3925375, "mato grosso"],
        'brasilia':[-10.3333333, -53.2, "distrito federal"]
    }

def format_name(name: str):
    """
    Format name to remove punctuation and spaces
    """
    name = re.sub("([^a-zà-üA-ZÀ-Ü0-9])", " ", name)
    nfkd_str = normalize("NFKD", name)
    name = "".join(
        [c for _, c in enumerate(nfkd_str) if not combining(c)]
    ).lower()
    while "  " in name:
        name = name.replace("   ", " ")
        name = name.replace("  ", " ")
    return name

def preprocess_dataset(df: pd.DataFrame, year: int):
    occ = df['NO_MUNICIPIO_PROVA'].value_counts().values
    ctt = df['NO_MUNICIPIO_PROVA'].value_counts().index.tolist()

    cor_brancos = []
    cor_pretos = []
    ibge_id = []
    
    for city in ctt:
        df_temp = df[df["NO_MUNICIPIO_PROVA"]==city].reset_index(drop=True)
        value_counts = df_temp["TP_COR_RACA"].value_counts()
        try:
            cor_brancos.append(value_counts[1])
        except:
            cor_brancos.append(0)
        try:
            cor_pretos.append(value_counts[2])
        except:
            cor_pretos.append(0)
        ibge_id.append(df_temp["CO_MUNICIPIO_PROVA"][0])

    df = pd.DataFrame({
        'Ano':year,
        'Num_Candidatos':occ,
        'Num_Brancos':cor_brancos,
        'Num_Pretos':cor_pretos,
        'id':ibge_id,
        'Cidade':ctt
    })
    latitude = []
    longitude = []
    states = []
    
    for i,row in df.iterrows():
        city = format_name(row['Cidade'])
        df.loc[i,'Cidade'] = city
        
        if city not in list(cities.keys()):
            location = geolocator.geocode(str(city+',Brazil'))
            match = geolocator.reverse(location.raw['lat']+","+location.raw['lon'])
            address = match.raw['address']
            state = format_name(address.get('state',''))
            
            cities[city] = [location.raw['lat'], location.raw['lon'], state]
            
            latitude.append(location.raw['lat'])
            longitude.append(location.raw['lon'])
            states.append(state)
        else:
            latitude.append(cities[city][0])
            longitude.append(cities[city][1])
            states.append(cities[city][2])
    
    df['Estado'] = states
    df['Latitude'] = latitude
    df['Longitude'] = longitude
    
    return df

## Carregamento dos dados e pré-processamento

# 01
enem13 = pd.read_csv(f"{PATH}/microdados_enem_2013/DADOS/MICRODADOS_ENEM_2013.csv", sep=';', encoding = "ISO-8859-1")[cols]
enem13 = preprocess_dataset(enem13, 2013)
enem13

# 02
enem14 = pd.read_csv(f"{PATH}/microdados_enem_2014/DADOS/MICRODADOS_ENEM_2014.csv", sep=';', encoding = "ISO-8859-1")[cols]
enem14 = preprocess_dataset(enem14, 2014)
enem14

# 03
enem15 = pd.read_csv(f"{PATH}/microdados_enem_2015/DADOS/MICRODADOS_ENEM_2015.csv", sep=';', encoding = "ISO-8859-1")[cols]
enem15 = preprocess_dataset(enem15, 2015)
enem15

# 04
enem16 = pd.read_csv(f"{PATH}/microdados_enem_2016/DADOS/MICRODADOS_ENEM_2016.csv", sep=';', encoding = "ISO-8859-1")[cols]
enem16 = preprocess_dataset(enem16, 2016)
enem16

# 05
enem17 = pd.read_csv(f"{PATH}/microdados_enem_2017/DADOS/MICRODADOS_ENEM_2017.csv", sep=';', encoding = "ISO-8859-1")[cols]
enem17 = preprocess_dataset(enem17, 2017)
enem17

# 06
enem18 = pd.read_csv(f"{PATH}/microdados_enem_2018/DADOS/MICRODADOS_ENEM_2018.csv", sep=';', encoding = "ISO-8859-1")[cols]
enem18 = preprocess_dataset(enem18, 2018)
enem18

# 07
enem19 = pd.read_csv(f"{PATH}/microdados_enem_2019/DADOS/MICRODADOS_ENEM_2019.csv", sep=';', encoding = "ISO-8859-1")[cols]
enem19 = preprocess_dataset(enem19, 2019)
enem19

# 08
enem20 = pd.read_csv(f"{PATH}/microdados_enem_2020/DADOS/MICRODADOS_ENEM_2020.csv", sep=';', encoding = "ISO-8859-1")[cols]
enem20 = preprocess_dataset(enem20, 2020)
enem20

# 09
enem21 = pd.read_csv(f"{PATH}/microdados_enem_2021/DADOS/MICRODADOS_ENEM_2021.csv", sep=';', encoding = "ISO-8859-1")[cols]
enem21 = preprocess_dataset(enem21, 2021)
enem21

# 10
enem22 = pd.read_csv(f"{PATH}/microdados_enem_2022/DADOS/MICRODADOS_ENEM_2022.csv", sep=';', encoding = "ISO-8859-1")[cols]
enem22 = preprocess_dataset(enem22, 2022)
enem22

## Finalização da base de dados final

if os.path.exists(path) == False:
    file = open(path, 'wb')
    pickle.dump(cities, file)
    file.close()

final = pd.concat([enem13,enem14,enem15,enem16,enem17,enem18,enem19,enem20,enem21,enem22]).reset_index(drop=True)
final[["Latitude", "Longitude"]] = final[["Latitude", "Longitude"]].apply(pd.to_numeric)

final.to_pickle(f"{PATH}/enem_data.pickle")
final.to_csv(f"{PATH}/enem_data.csv")

final.info()

final

print("Elapsed time (h:mm:ss):",str(timedelta(seconds=(time.time() - start))))
