==> notebooks/03_Visualizacao02.ipynb <==
import pandas as pd
import plotly.express as px
px.set_mapbox_access_token("pk.eyJ1IjoibHVpemhlbWVsbyIsImEiOiJjbG9hZGkzdDQwaG94Mm1udmg4Mmx0ZG45In0.etjGdsIyRFv-oaU74-8DTg")

df = pd.read_pickle('../../dados_enem/enem_data.pickle')
df

fig = px.scatter_mapbox(
    data_frame=df,
    lat="Latitude",
    lon="Longitude",
    size="Num_Candidatos",
    animation_frame="Ano",
    text=None,
    color="Estado",
    color_continuous_scale=px.colors.cyclical.IceFire,
    size_max=40,
    zoom=3,
    mapbox_style="carto-positron",
    height=700,
    center = dict(lat = -15.685413, lon = -53.679588),
    labels={
        "Num_Candidatos": "Número de candidatos"
    }
)

#fig.show()
fig.write_html("../src/vis/vis02.html")
