==> notebooks/04_Visualizacao03.ipynb <==
import json
from numpy import round
import pandas as pd
import plotly.express as px
px.set_mapbox_access_token("pk.eyJ1IjoibHVpemhlbWVsbyIsImEiOiJjbG9hZGkzdDQwaG94Mm1udmg4Mmx0ZG45In0.etjGdsIyRFv-oaU74-8DTg")

## Carregamento dos dados

df = pd.read_pickle('../../dados_enem/enem_data.pickle')
df["id"] = df["id"].astype("str")

df["Num_Pretos"] = df["Num_Pretos"].replace(0, 1)
df["Relacao_Brancos_Pretos"] = round(df["Num_Brancos"].values/df["Num_Pretos"].values, 3)
df

geo_cidades = json.load(open('../data//geojs-100-mun.json',))
geo_cidades['features'][0]

## Geração da visualização

fig = px.choropleth_mapbox(
    data_frame=df, 
    geojson=geo_cidades, 
    locations='id', 
    featureidkey='properties.id',
    color='Relacao_Brancos_Pretos',
    animation_frame="Ano",
    color_continuous_scale='Pinkyl',
    range_color=(0,15),
    mapbox_style='dark',
    zoom=3,
    center = {"lat":-15.685413,"lon":-53.679588},
    opacity=0.7,
    height=700,
    labels={
        "id": "Código do município",
        "Relacao_Brancos_Pretos": "Relação Brancos/Pretos"
    }
)
fig.update_layout(margin={'r':0,'t':0,'l':0, 'b':0})
fig.update_traces(marker_line_width=0.01)

#fig.show()
fig.write_html("../src/vis/vis03.html")
